Move `IconButton` to `Icons.kt`.
